Add another Type X+ board variant
The Type X+ 013A is used in Battle Gear 4 Tuned pro. It is pretty much identical to the 005 and 005B that is used in SD cabinets except it uses a more powerful 9800PRO due to the higher resolution (Tuned in pro mode runs at 1366x768 instead of 800x600). Unfortunately dont have the HDD label for tuned pro though. Also updated info for the 005A variant which ive seen in european BG4 cabinets.